app/infrastructure/image_utils.py
from PIL import Image
import torchvision.transforms as transforms

transform = transforms.Compose([
    transforms.Resize((224, 224)),
    transforms.ToTensor()
])

def preprocess_image(image: Image.Image):
    return transform(image).unsqueeze(0)